Cleanup: Move AsyncEngine to libbase.la
