+++ src/Makefile.am
@@ -195,8 +195,6 @@ squid_SOURCES = \
 	$(WINSVC_SOURCE) \
 	AccessLogEntry.cc \
 	AccessLogEntry.h \
-	AsyncEngine.cc \
-	AsyncEngine.h \
 	AuthReg.h \
 	BodyPipe.cc \
 	BodyPipe.h \
