+++ src/base/Makefile.am
@@ -20,6 +20,7 @@ libbase_la_SOURCES = \
 	AsyncCallQueue.h \
 	AsyncCallbacks.h \
 	AsyncCbdataCalls.h \
+	AsyncEngine.h \
 	AsyncFunCalls.h \
 	AsyncJob.cc \
 	AsyncJob.h \
